+++ scripts/tests.sh
@@ -83,9 +83,6 @@ done
 printTask "Testing Python scripts..."
 "$REPO_ROOT/test/pyscriptTests.py"
 
-printTask "Testing LSP..."
-"$REPO_ROOT/test/lsp.py" "${SOLIDITY_BUILD_DIR}/solc/solc"
-
 printTask "Running commandline tests..."
 # Only run in parallel if this is run on CI infrastructure
 if [[ -n "$CI" ]]